detect/requirements.txt
psycopg2-binary
requests
Flask
python-dotenv
pytz
